Fixed compiler warnings for x64 and x86. Aligned allocation of SSE arrays.

+++ source/CmdLineInterface/CmdLineInterface.cpp
@@ -16,7 +16,7 @@ CmdLineInterface::~CmdLineInterface() {
 }
 
 void CmdLineInterface::addTokens(const char* str) {
-	int len = strlen(str)+1;
+	int len = (int)strlen(str)+1;
 	char* tmp = new char[len];
 	char* ptr = tmp;
 	char* token_start = tmp;
